+++ parva_os/src/main.rs
@@ -9,7 +9,6 @@ extern crate alloc;
 use parva_os::println;
 use core::panic::PanicInfo;  // We import this to get information about future panics
 use bootloader::{entry_point, BootInfo};
-use alloc::{boxed::Box, rc::Rc, vec, vec::Vec};
 
 entry_point!(kernel_main);
 
@@ -18,7 +17,10 @@ fn kernel_main(boot_info: &'static BootInfo) -> ! {
     use parva_os::memory::{self, BootInfoFrameAllocator};
     use x86_64::VirtAddr;
 
-    println!("Hello World{}", "!");
+    println!("Welcome to ParvaOS");
+    println!("Copyright (c) 2025 Francesco Giannice");
+    println!("v. 0.0.1\n");
+
     parva_os::init();
 
     let phys_mem_offset = VirtAddr::new(boot_info.physical_memory_offset);
@@ -27,34 +29,11 @@ fn kernel_main(boot_info: &'static BootInfo) -> ! {
 
     allocator::init_heap(&mut mapper, &mut frame_allocator).expect("heap initialization failed");
 
-    // allocate a number on the heap
-    let heap_value = Box::new(41);
-    println!("heap_value at {:p}", heap_value);
-
-    // create a dynamically sized vector
-    let mut vec = Vec::new();
-    for i in 0..500 {
-        vec.push(i);
-    }
-    println!("vec at {:p}", vec.as_slice());
-
-    // create a reference counted vector -> will be freed when count reaches 0
-    let reference_counted = Rc::new(vec![1, 2, 3]);
-    let cloned_reference = reference_counted.clone();
-    println!(
-        "current reference count is {}",
-        Rc::strong_count(&cloned_reference)
-    );
-    core::mem::drop(reference_counted);
-    println!(
-        "reference count is {} now",
-        Rc::strong_count(&cloned_reference)
-    );
-
+    println!("\nPress 'Enter' to START\n");
+    
     #[cfg(test)]
     test_main();
 
-    println!("It did not crash!");
     parva_os::hlt_loop();
 }