+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **CPU Exceptions**: added a way to manage CPU exception without crashing the entire system
 - **Paging**: implemented virtual memory management by introducing paging, enabling memory isolation and efficient use of physical memory.
 - **Heap allocation**: introduced a dynamic memory allocator to manage memory allocation and deallocation at runtime, enabling more flexible and efficient use of system resources.
+- **Shell**: added a basic shell with a cursor, backspace working an `hello` as example command
 
 ## 0.0.0 (2025-04-24)
 - Started ParvaOS project